+++ index.php
@@ -10,7 +10,7 @@
 
     // core modules
     require __DIR__ . '/src/Core/validation.php';
-    require __DIR__ . '/src/Core/customdate.php';
+    require __DIR__ . '/src/Core/leafdate.php';
 
     // helpers    
     require __DIR__ . '/src/Helpers/constants.php';
@@ -29,12 +29,18 @@
     $leaf = new Leaf\Core\Leaf;
     $response = new Leaf\Core\Response;
     $request = new Leaf\Core\Request;
+    $date = new Leaf\Core\LeafDate;
 
     $leaf->get('/home', function() use($response, $request) {
         $id = $request->getParam('id');
         $body = $request->getBody();
         // echo $response->respond(["message" => "Welcome to the Leaf Framework....your id is ".$id ]);
         echo json_encode($body);
+    });
+
+    $leaf->get('/date', function() use($date, $response) {
+        $data = $date->GetDayFromNumber(2);
+        $response->respond(["message" => $data]);
     });
 
 	$leaf->run();