+++ src/core/response.php
@@ -4,19 +4,19 @@
     class Response {
         public function respond($data) {
             header('Content-Type: application/json');
-            return json_encode($data);
+            echo json_encode($data);
         }
 
         public function respondWithCode($data, $code = 200) {
             header('Content-Type: application/json');
             $dataToPrint = array('data' => $data, 'code' => $code);
-            return json_encode($dataToPrint);
+            echo json_encode($dataToPrint);
         }
 
         public function throwErr($error, $code) {
             header('Content-Type: application/json');
             $dataToPrint = array('error' => $error, 'code' => $code);
-            return json_encode($dataToPrint);
+            echo json_encode($dataToPrint);
         }
 
         // has no use at the moment
@@ -26,11 +26,11 @@
 
         public function renderHtmlPage($file) {
             header('Content-Type: text/html');
-            return require $file;
+            echo require $file;
         }
 
         public function renderMarkup($markup) {
             header('Content-Type: text/html');
-            return $markup;
+            echo $markup;
         }
     };